# Use direct cache in index reader for symbol values