@bwplotka 2020-12-09
Nice. I will check tomorrow but I still think we could just adjust symbolFactor to be something different than 32. If not 1. 

This should also help with allocs, wonder what's the main source. I guess we could together Investigate profiles (: