@bwplotka 2020-12-09
Can you double check you are running this tests for more or equal than minute. Especially if something is taking long time, worth to rule out laptop resources saturation