@yeya24 2020-12-10
> Can you double check you are running this tests for more or equal than minute. Especially if something is taking long time, worth to rule out laptop resources saturation

The result above is under 30s bench time. I did another round of benchmarks with 1m bench time and the result is similar.

```
name                                                             old time/op    new time/op    delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E        14.7s ± 0%      8.3s ± 0%  -43.40%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E       3.50s ± 0%     3.32s ± 0%   -5.09%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E     2.55s ± 0%     1.14s ± 0%  -55.33%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E          15.5s ± 0%      6.6s ± 0%  -57.26%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E         2.22s ± 0%     2.02s ± 0%   -9.24%

name                                                             old alloc/op   new alloc/op   delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E       4.05GB ± 0%    3.50GB ± 0%  -13.54%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E      1.09GB ± 0%    1.08GB ± 0%   -0.38%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E     663MB ± 0%     545MB ± 0%  -17.78%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E         2.67GB ± 0%    2.22GB ± 0%  -16.99%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E         631MB ± 0%     631MB ± 0%   -0.00%

name                                                             old allocs/op  new allocs/op  delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E        38.7M ± 0%     15.5M ± 0%  -59.85%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E        412k ± 0%      229k ± 0%  -44.49%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E     5.85M ± 0%     2.10M ± 0%  -64.20%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E          30.4M ± 0%      7.2M ± 0%  -76.31%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E         38.5k ± 0%     38.4k ± 0%   -0.10%
```