@bwplotka 2020-12-10
Awesome, thanks for checking, I want to try something additional before we move to this 🤗