@bwplotka 2020-12-10
I was thinking about something like this: https://github.com/thanos-io/thanos/pull/3561

Will compare the results from your PR as well (: