@bwplotka 2020-12-10
Interesting, looks like caching is better:

```
![image](https://user-images.githubusercontent.com/6950331/101775395-53a34f00-3ae7-11eb-8eb3-0a5ebc2d2af5.png)

```