@csmarchbanks 2020-12-10
I just want to chime in and say it is awesome to see this optimization helping out other parts of Thanos too!